Sacrifice specificity for sensitivity in status view branch pattern
The idea here is not to validate a branch's name, but to parse it from the
git status output. Git will not allow a branch to be created with an
invalid name. Thus what I've tried to do is to create a simple regex that
will recognize any valid branch, while not worrying about each branch
naming rule at http://git-scm.com/docs/git-check-ref-format#_description.